Algorithm to find words in ospd file
Finds words in the ospd file that can contain the character on the board, and are made of only letters in the char array. I can't get the code to compile. We can change this so that it calculates the point value and stores it in another array, and also so that only the top three words are stored.